+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/car/search/PlaceSearchScreen.kt
@@ -9,6 +9,7 @@ import androidx.car.app.model.ItemList
 import androidx.car.app.model.Row
 import androidx.car.app.model.SearchTemplate
 import androidx.car.app.model.Template
+import androidx.lifecycle.lifecycleScope
 import com.mapbox.androidauto.R
 import com.mapbox.androidauto.car.feedback.core.CarFeedbackSearchOptions
 import com.mapbox.androidauto.car.feedback.core.CarFeedbackSender
@@ -18,18 +19,27 @@ import com.mapbox.androidauto.car.preview.CarRouteRequestCallback
 import com.mapbox.androidauto.car.preview.RoutePreviewCarContext
 import com.mapbox.androidauto.internal.logAndroidAuto
 import com.mapbox.androidauto.internal.logAndroidAutoFailure
+import com.mapbox.maps.MapboxExperimental
+import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.base.route.NavigationRoute
+import com.mapbox.navigation.core.internal.extensions.attachCreated
 import com.mapbox.search.result.SearchSuggestion
+import kotlinx.coroutines.launch
 
 /**
  * This screen allows the user to search for a destination.
  */
-class SearchScreen(
+@OptIn(MapboxExperimental::class, ExperimentalPreviewMapboxNavigationAPI::class)
+class PlaceSearchScreen(
     private val searchCarContext: SearchCarContext,
 ) : Screen(searchCarContext.carContext) {
 
     @VisibleForTesting
-    var itemList = buildErrorItemList(R.string.car_search_no_results)
+    internal var itemList = buildNoItemsList(R.string.car_search_no_results)
+        private set(value) {
+            field = value
+            invalidate()
+        }
 
     // Cached to send to feedback.
     private var searchSuggestions: List<SearchSuggestion> = emptyList()
@@ -43,16 +53,20 @@ class SearchScreen(
         }
 
         override fun onUnknownCurrentLocation() {
-            onErrorItemList(R.string.car_search_unknown_current_location)
+            itemList = buildNoItemsList(R.string.car_search_unknown_current_location)
         }
 
         override fun onDestinationLocationUnknown() {
-            onErrorItemList(R.string.car_search_unknown_search_location)
+            itemList = buildNoItemsList(R.string.car_search_unknown_search_location)
         }
 
         override fun onNoRoutesFound() {
-            onErrorItemList(R.string.car_search_no_results)
+            itemList = buildNoItemsList(R.string.car_search_no_results)
         }
+    }
+
+    init {
+        attachCreated(searchCarContext.carPlaceSearch)
     }
 
     override fun onGetTemplate(): Template {
@@ -77,7 +91,7 @@ class SearchScreen(
                                 .getSearchFeedbackPoll(searchCarContext.carContext),
                         ) {
                             CarFeedbackSearchOptions(searchSuggestions = searchSuggestions)
-                        }.getAction(this@SearchScreen)
+                        }.getAction(this@PlaceSearchScreen)
                     )
                     .build()
             )
@@ -86,18 +100,20 @@ class SearchScreen(
             .build()
     }
 
-    fun doSearch(searchText: String) {
-        searchCarContext.carSearchEngine.search(searchText) { suggestions ->
+    @VisibleForTesting
+    internal fun doSearch(searchText: String) {
+        lifecycleScope.launch {
+            val suggestions = searchCarContext.carPlaceSearch.search(searchText)
+                .getOrDefault(emptyList())
             searchSuggestions = suggestions
-            if (suggestions.isEmpty()) {
-                onErrorItemList(R.string.car_search_no_results)
+            itemList = if (suggestions.isEmpty()) {
+                buildNoItemsList(R.string.car_search_no_results)
             } else {
                 val builder = ItemList.Builder()
                 suggestions.forEach { suggestion ->
                     builder.addItem(searchItemRow(suggestion))
                 }
-                itemList = builder.build()
-                invalidate()
+                builder.build()
             }
         }
     }
@@ -115,7 +131,9 @@ class SearchScreen(
 
     private fun onClickSearch(searchSuggestion: SearchSuggestion) {
         logAndroidAuto("onClickSearch $searchSuggestion")
-        searchCarContext.carSearchEngine.select(searchSuggestion) { searchResults ->
+        lifecycleScope.launch {
+            val searchResults = searchCarContext.carPlaceSearch.select(searchSuggestion)
+                .getOrDefault(emptyList())
             logAndroidAuto("onClickSearch select ${searchResults.joinToString()}")
             if (searchResults.isNotEmpty()) {
                 searchCarContext.carRouteRequest.request(
@@ -126,19 +144,7 @@ class SearchScreen(
         }
     }
 
-    private fun onErrorItemList(@StringRes stringRes: Int) {
-        itemList = buildErrorItemList(stringRes)
-        invalidate()
-    }
-
-    private fun buildErrorItemList(@StringRes stringRes: Int) = ItemList.Builder()
+    private fun buildNoItemsList(@StringRes stringRes: Int) = ItemList.Builder()
         .setNoItemsMessage(carContext.getString(stringRes))
         .build()
-
-    companion object {
-        // TODO turn this into something typesafe
-        fun parseResult(results: Any?): Any? {
-            return results
-        }
-    }
 }