+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/car/search/SearchCarContext.kt
@@ -3,12 +3,15 @@ package com.mapbox.androidauto.car.search
 import com.mapbox.androidauto.MapboxCarApp
 import com.mapbox.androidauto.car.MainCarContext
 import com.mapbox.androidauto.car.preview.CarRouteRequest
+import com.mapbox.androidauto.internal.car.search.CarPlaceSearch
+import com.mapbox.androidauto.internal.car.search.CarPlaceSearchImpl
+import com.mapbox.androidauto.internal.car.search.CarSearchLocationProvider
 
 /**
  * Contains the dependencies for the search feature.
  */
 class SearchCarContext(
-    val mainCarContext: MainCarContext
+    val mainCarContext: MainCarContext,
 ) {
     /** MainCarContext **/
     val carContext = mainCarContext.carContext
@@ -16,9 +19,9 @@ class SearchCarContext(
     val feedbackPollProvider = mainCarContext.feedbackPollProvider
 
     /** SearchCarContext **/
-    val carSearchEngine = CarSearchEngine(
-        mainCarContext.searchEngine,
-        MapboxCarApp.carAppLocationService().navigationLocationProvider
+    internal val carPlaceSearch: CarPlaceSearch = CarPlaceSearchImpl(
+        mainCarContext.carPlaceSearchOptions,
+        CarSearchLocationProvider()
     )
     val carRouteRequest = CarRouteRequest(
         mainCarContext.mapboxNavigation,