+++ qa-test-app/src/main/java/com/mapbox/navigation/qa_test_app/QaTestApplication.kt
@@ -2,7 +2,6 @@ package com.mapbox.navigation.qa_test_app
 
 import android.app.Application
 import com.mapbox.androidauto.MapboxCarApp
-import com.mapbox.navigation.qa_test_app.car.search.MapboxCarSearchApp
 
 class QaTestApplication : Application() {
 
@@ -10,6 +9,5 @@ class QaTestApplication : Application() {
         super.onCreate()
 
         MapboxCarApp.setup(this)
-        MapboxCarSearchApp.setup()
     }
 }